OpenCV  for MATH/requirements.txt
opencv-python==4.8.1.78
cvzone==1.6.1
mediapipe==0.10.21
numpy==1.26.4
google-generativeai
Pillow
streamlit 